Update folders structure
