+++ src/features/Tasks/queries.ts
@@ -1,25 +1,31 @@
-import { useQuery } from '@tanstack/react-query';
 import AsyncStorage from '@react-native-async-storage/async-storage';
-import { Tasks } from './types';
+import { useQuery } from '@tanstack/react-query';
+import { Tasks } from '../../api/types';
+
+const taskKeys = {
+    all: ['tasks'] as const,
+    list: () => [...taskKeys.all, 'list'] as const,
+};
 
 const useTasks = () => {
     return useQuery<Tasks>({
-        queryKey: ['tasks'],
+        queryKey: taskKeys.list(),
         queryFn: async () => {
+            let allTasks = [] as Tasks;
+
             try {
                 const value = await AsyncStorage.getItem('@all_tasks');
 
                 if (value !== null) {
-                    return JSON.parse(value);
+                    allTasks = JSON.parse(value);
                 }
-
-                return [];
             } catch (e) {
                 console.log(e);
-                return [];
             }
+
+            return allTasks;
         },
     });
 };
 
-export { useTasks };
+export { taskKeys, useTasks };
